// types.ts
export interface Project {
  title: string;
  description: string;
  features: string[];
  tech: string[];
  icon: string;
}

export interface Experience {
  role: string;
  company: string;
  period: string;
  highlights: string[];
}

export interface SkillGroup {
  category: string;
  skills: string[];
}